Micah Crawford
Hi, my name is Micah Crawford. I am technically a senior at ETSU, but due to my transfer I will become a second year senior. What I am interested in learning is how to cut down on code and get a better understanding about it. Java was not a pleasant  experience for me so hopefully C# will be better. I am a computer science major with a concentration in IT. I really only have one hobby and that would be photography, if you want to view some of my work you can find my page on instagram and facebook @micah.alexander.photos